move handlePullRequestEdited into owl-bot.ts
